Changed encryption mode to Signed
This is to prevent an attack where an intruder encrypts a phony file pretending to be part of the group, which is possible with Anonymous mode based on a pair of dummy private and public keys. Now the private key must belong to one of the listed users, or decryption will fail. To allow decryption of files encrypted by a former user, this user can be deactivated for encryption by preceding his/her name by a '$' character; legacy users are not active for encryption, and their public keys are used if no matching public key for an active user is found. A salt was added for the key-stretching process, contained as constant GroupName in GroupKeys.js